chapter-2/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1


# backwards compatibility
pathlib2

# Added
jupyter
pandas
matplotlib
numpy
sklearn
requests

# dev
black
